Add interaction tracking and limit team registration button interactions. Fixed a bug in the join team button handler.
Fixed a bug in DynamicGrid where is was always creating the columns with 1 row on construction. This has been changed to generate the correct initial amount as given to the constructor.

Updated Button, SelectComponent and TextInput component names to read better.

Removed a neeless check when creating a Message's builder regarding adding content.
Message no longer accepts empty component rows.
Message now correctly validates SelectCoponents when building.

Fixed an issue in MessageExtensions where adding content would always be prefixed with a new line, even if it is the only content after adding.

Updated ButtonFactory.CloseButton to not be static to allow for DI to potentially function in the future. The Create method is non-static so it also did not make sense for the CloseButton to be as it means the class is referenced differently for no good reason.

Added InteractionsTracker and InteractionTracker to track a DiscordInteraction. This allows validators to check if a user has an active instance and can deny them springing more.

Moved the Conclude method from IInteractionHandler to InteractionTracker as the handler is meant to only exist until the request has be handled and only the pipleline should ever have access to a handler.

ConcludeInteractionButton request will now accept messages to delete and can limit interactions to a specific user so that only the user involved in the interaction can conclude it.

Removed RequestsTracker.ActiveCount as it is now on the InteractionsTracker and is not needed here.

In Validator, renamed ActiveRequestInstances to ActiveInteractions and updated it to work with InteractionsTracker.
Added UserInteraction validation check to limit the interaction to a specific user.
Added GetTooManyInteractionInstancesError to allow for a common error message to send to the user.

CreateTeamButtonValidator and CreateTeamModalValidator now only allow one interaction to be active per user and for their request interactions to be mutually exclusive.

CreateTeamModalValidator now checks if the user is not on a team.
Removed the semaphores from CreateTeamModalValidator their use is now handled better through interaction limiting.

CreateTeamModalHandler now concludes the interaction when it has finished processing.

JoinTeamButtonHandler now has a conclude interaction button since the validator now limits interactions.
It no longer sends the response as ephemeral since it has a conclude interaction button thus needs to ensure this is the only way the user can delete the message so that the interaction can be concluded and the user not locked out of further interactions.
Since the response is not ephemeral, the message content has been updated to mention the user that is interacting with it so they know which instance is theirs.
Fixed an issue where it would send a SelectComponent with no options if there were no teams to select from. The Discord API would throw, thus just a message and no select is sent in this case.

Removed the semaphores from CreateTeamModalValidator as they are no longer needed since the interactions are limited via the button.
Added a UserNotOnATeam check just in case the were added to a team since pressing the button to open the modal.