Add IRequestResponse, update SubmitDropButtonHandler  and fix BingoValidator image type validation.
Added IRequestResponse to be used on IRequests (not directly) to notify the MediatRExtensions the request has a response to send to Discord. Previously it was only interaction requests that could do this and that was restrictive.

Message now takes the DiscordChannel to send to. The message services does not.

Updated SubmitDropButtonHandler responses, request services use select options order.

Fixed BingoValidator image type validation.