Fix SubmitDropSubmitButtonHandler saving to db.
Removed dw from InteractionHandler as only classes that need it should create it to avoid unessassary connections to the db. Also, it can cause tracking issues if it's used and a sub request uses its own so it's better for the base request to create a dw then send it down the line to the sub requests.