Add ClassInitialize and bugfix such as TileValues row and column generation.
Add rows and columns to Grid<> for ease for accessing.

Added ClassInitialize to ScoringTests to ensure TileValues doesn't receive duplicate data.

Clean up and fixed a test.

Fixed a bug in TeamScore where it wasn't checking if the tile was completed.
Fixed a bug in TeamScore where it was retrieving the tilesValues each time it was looping over them.

Fixed row and column generation in TileValues and clean up.