Add IDataParser and TileSelectValidator.
The data parsers takes the input data given to the public TileSelectValidator.Validate and converts it to the format the TSV needs which is given to the protected abstract TileSelectValidator.Validate.

This allows tests to only need to test the input data on the parsers and not the TSVs, resulting in much less complication and many less tests.